Move PointsTo to MemoryModel.
