+++ include/DDA/DDAVFSolver.h
@@ -12,7 +12,7 @@
 #include "MSSA/SVFGBuilder.h"
 #include "WPA/Andersen.h"
 #include "Util/SCC.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 #include <algorithm>
 
 namespace SVF
