+++ include/MTA/FSMPTA.h
@@ -12,7 +12,7 @@
 #include "WPA/FlowSensitive.h"
 #include "MSSA/SVFGBuilder.h"
 #include "MTA/LockAnalysis.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 #include "MTA/MHP.h"
 
 namespace SVF
