+++ include/MemoryModel/PTAStat.h
@@ -31,7 +31,7 @@
 #define ANDERSENSTAT_H_
 
 #include "Util/BasicTypes.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 #include <iostream>
 #include <map>
 #include <string>
