+++ include/Util/TypeBasedHeapCloning.h
@@ -14,7 +14,6 @@
 #include "Graphs/PAG.h"
 #include "MemoryModel/PointerAnalysisImpl.h"
 #include "Util/BasicTypes.h"
-#include "Util/PointsTo.h"
 
 namespace SVF
 {
