+++ include/WPA/FlowSensitiveTBHC.h
@@ -15,7 +15,7 @@
 #include "MSSA/SVFGBuilder.h"
 #include "Util/TypeBasedHeapCloning.h"
 #include "WPA/FlowSensitive.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 
 namespace SVF
 {
