+++ lib/DDA/DDAPass.cpp
@@ -12,7 +12,6 @@
 #include "DDA/ContextDDA.h"
 #include "DDA/DDAClient.h"
 #include "SVF-FE/PAGBuilder.h"
-#include "Util/PointsTo.h"
 
 #include <sstream>
 #include <limits.h>
