+++ lib/DDA/DDAStat.cpp
@@ -9,7 +9,7 @@
 #include "DDA/FlowDDA.h"
 #include "DDA/ContextDDA.h"
 #include "Graphs/SVFGStat.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 
 #include <iomanip>
