+++ lib/MTA/MTAAnnotator.cpp
@@ -8,7 +8,7 @@
 #include "Util/Options.h"
 #include "MTA/MTAAnnotator.h"
 #include "MTA/LockAnalysis.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 #include <sstream>
 
 using namespace SVF;
