+++ lib/MemoryModel/PointerAnalysisImpl.cpp
@@ -9,7 +9,6 @@
 #include "MemoryModel/PointerAnalysisImpl.h"
 #include "SVF-FE/CPPUtil.h"
 #include "SVF-FE/DCHG.h"
-#include "Util/PointsTo.h"
 #include "Util/Options.h"
 #include "Util/IRAnnotator.h"
 #include <fstream>
