+++ lib/Util/PTAStat.cpp
@@ -32,7 +32,6 @@
 #include "MemoryModel/PTAStat.h"
 #include "MemoryModel/PointerAnalysisImpl.h"
 #include "Graphs/PAG.h"
-#include "Util/PointsTo.h"
 #include "Util/Options.h"
 
 using namespace SVF;
