+++ lib/WPA/Andersen.cpp
@@ -29,7 +29,7 @@
 
 #include "Util/Options.h"
 #include "SVF-FE/LLVMUtil.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 #include "Util/Options.h"
 #include "WPA/Andersen.h"
 #include "WPA/Steensgaard.h"
