+++ lib/WPA/AndersenSCD.cpp
@@ -28,7 +28,7 @@
  */
 
 #include "WPA/AndersenSFR.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 
 using namespace SVF;
 using namespace SVFUtil;
