+++ lib/WPA/AndersenStat.cpp
@@ -31,7 +31,6 @@
 #include "MemoryModel/PointerAnalysis.h"
 #include "WPA/WPAStat.h"
 #include "WPA/Andersen.h"
-#include "Util/PointsTo.h"
 
 using namespace SVF;
 using namespace SVFUtil;
