+++ lib/WPA/AndersenWaveDiff.cpp
@@ -28,7 +28,7 @@
  */
 
 #include "WPA/Andersen.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 
 using namespace SVF;
 using namespace SVFUtil;
