+++ lib/WPA/FlowSensitive.cpp
@@ -34,7 +34,7 @@
 #include "WPA/WPAStat.h"
 #include "WPA/FlowSensitive.h"
 #include "WPA/Andersen.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 
 
 using namespace SVF;
