+++ lib/WPA/FlowSensitiveTBHC.cpp
@@ -13,7 +13,7 @@
 #include "WPA/FlowSensitiveTBHC.h"
 #include "WPA/WPAStat.h"
 #include "WPA/Andersen.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 
 using namespace SVF;
