+++ lib/WPA/VersionedFlowSensitive.cpp
@@ -10,7 +10,7 @@
 #include "WPA/Andersen.h"
 #include "WPA/VersionedFlowSensitive.h"
 #include "Util/Options.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 #include <iostream>
 
 using namespace SVF;
