+++ lib/WPA/VersionedFlowSensitiveStat.cpp
@@ -10,7 +10,7 @@
 #include "SVF-FE/LLVMUtil.h"
 #include "WPA/WPAStat.h"
 #include "WPA/VersionedFlowSensitive.h"
-#include "Util/PointsTo.h"
+#include "MemoryModel/PointsTo.h"
 
 using namespace SVF;
 using namespace SVFUtil;
